# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: "Generated Application API"
  version: "1.0.0"
  description: |
    API generated by RNC from legacy source code.
    Business rules are preserved from the original system.
servers:
  - url: http://localhost:3059/api
    description: Local development
paths:
...
  /programa_socials:
    get:
      summary: List ProgramaSocials
      operationId: listProgramaSocials
      tags: [ProgramaSocial]
      parameters:
        - name: page
          in: query
          schema: {type: integer, default: 1}
        - name: limit
          in: query
          schema: {type: integer, default: 20, maximum: 100}
      responses:
        "200":
          description: Paginated list of ProgramaSocials
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ProgramaSocialList"}
    post:
      summary: Create ProgramaSocial
      operationId: createProgramaSocial
      tags: [ProgramaSocial]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ProgramaSocialInput"}
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ProgramaSocial"}
        "400":
          description: Validation error
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ValidationError"}
  "/programa_socials/{id}":
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: integer}
    get:
      summary: Get ProgramaSocial by id
      operationId: getProgramaSocialById
      tags: [ProgramaSocial]
      responses:
        "200":
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ProgramaSocial"}
        "404": {description: Not found}
    put:
      summary: Update ProgramaSocial
      operationId: updateProgramaSocial
      tags: [ProgramaSocial]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ProgramaSocialInput"}
      responses:
        "200":
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ProgramaSocial"}
        "400": {description: Validation error}
        "404": {description: Not found}
    delete:
      summary: Delete ProgramaSocial
      operationId: deleteProgramaSocial
      tags: [ProgramaSocial]
      responses:
        "204": {description: Deleted}
        "404": {description: Not found}
  /programa_social_grp_faixa_calculos:
    get:
      summary: List ProgramaSocialGrpFaixaCalculos
      operationId: listProgramaSocialGrpFaixaCalculos
      tags: [ProgramaSocialGrpFaixaCalculo]
      parameters:
        - name: page
          in: query
          schema: {type: integer, default: 1}
        - name: limit
          in: query
          schema: {type: integer, default: 20, maximum: 100}
      responses:
        "200":
          description: Paginated list of ProgramaSocialGrpFaixaCalculos
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ProgramaSocialGrpFaixaCalculoList"}
    post:
      summary: Create ProgramaSocialGrpFaixaCalculo
      operationId: createProgramaSocialGrpFaixaCalculo
      tags: [ProgramaSocialGrpFaixaCalculo]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ProgramaSocialGrpFaixaCalculoInput"}
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ProgramaSocialGrpFaixaCalculo"}
        "400":
          description: Validation error
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ValidationError"}
  "/programa_social_grp_faixa_calculos/{id}":
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: integer}
    get:
      summary: Get ProgramaSocialGrpFaixaCalculo by id
      operationId: getProgramaSocialGrpFaixaCalculoById
      tags: [ProgramaSocialGrpFaixaCalculo]
      responses:
        "200":
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ProgramaSocialGrpFaixaCalculo"}
        "404": {description: Not found}
    put:
      summary: Update ProgramaSocialGrpFaixaCalculo
      operationId: updateProgramaSocialGrpFaixaCalculo
      tags: [ProgramaSocialGrpFaixaCalculo]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ProgramaSocialGrpFaixaCalculoInput"}
      responses:
        "200":
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ProgramaSocialGrpFaixaCalculo"}
        "400": {description: Validation error}
        "404": {description: Not found}
    delete:
      summary: Delete ProgramaSocialGrpFaixaCalculo
      operationId: deleteProgramaSocialGrpFaixaCalculo
      tags: [ProgramaSocialGrpFaixaCalculo]
      responses:
        "204": {description: Deleted}
        "404": {description: Not found}
  /programa_social_grp_param_regionals:
    get:
      summary: List ProgramaSocialGrpParamRegionals
      operationId: listProgramaSocialGrpParamRegionals
      tags: [ProgramaSocialGrpParamRegional]
      parameters:
        - name: page
          in: query
          schema: {type: integer, default: 1}
        - name: limit
          in: query
          schema: {type: integer, default: 20, maximum: 100}
      responses:
        "200":
          description: Paginated list of ProgramaSocialGrpParamRegionals
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ProgramaSocialGrpParamRegionalList"}
    post:
      summary: Create ProgramaSocialGrpParamRegional
      operationId: createProgramaSocialGrpParamRegional
      tags: [ProgramaSocialGrpParamRegional]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ProgramaSocialGrpParamRegionalInput"}
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ProgramaSocialGrpParamRegional"}
        "400":
          description: Validation error
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ValidationError"}
  "/programa_social_grp_param_regionals/{id}":
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: integer}
    get:
      summary: Get ProgramaSocialGrpParamRegional by id
      operationId: getProgramaSocialGrpParamRegionalById
      tags: [ProgramaSocialGrpParamRegional]
      responses:
        "200":
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ProgramaSocialGrpParamRegional"}
        "404": {description: Not found}
    put:
      summary: Update ProgramaSocialGrpParamRegional
      operationId: updateProgramaSocialGrpParamRegional
      tags: [ProgramaSocialGrpParamRegional]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ProgramaSocialGrpParamRegionalInput"}
      responses:
        "200":
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ProgramaSocialGrpParamRegional"}
        "400": {description: Validation error}
        "404": {description: Not found}
    delete:
      summary: Delete ProgramaSocialGrpParamRegional
      operationId: deleteProgramaSocialGrpParamRegional
      tags: [ProgramaSocialGrpParamRegional]
      responses:
        "204": {description: Deleted}
        "404": {description: Not found}
...
components:
  schemas:
...
    ValidationError:
      type: object
      properties:
        errors:
          type: object
          additionalProperties:
            type: array
            items: {type: string}
...
    ProgramaSocial:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        codPrograma:
          type: string
          nullable: true
          description: "Cod Programa"
        nomePrograma:
          type: string
          nullable: true
          description: "Nome Programa"
        siglaPrograma:
          type: string
          nullable: true
          description: "Sigla Programa"
        tipoPrograma:
          type: string
          nullable: true
          description: "Tipo Programa"
        orgaoResponsavel:
          type: string
          nullable: true
          description: "Orgao Responsavel"
        leiCriacao:
          type: string
          nullable: true
          description: "Lei Criacao"
        dtCriacao:
          type: string
          nullable: true
          description: "Dt Criacao"
        dtEncerramento:
          type: string
          nullable: true
          description: "Dt Encerramento"
        sitPrograma:
          type: string
          nullable: true
          description: "Sit Programa"
        vlrBaseIndividual:
          type: string
          nullable: true
          description: "Vlr Base Individual"
        vlrBaseFamiliar:
          type: string
          nullable: true
          description: "Vlr Base Familiar"
        vlrTetoBenef:
          type: string
          nullable: true
          description: "Vlr Teto Benef"
        vlrPisoBenef:
          type: string
          nullable: true
          description: "Vlr Piso Benef"
        pctReajusteAnual:
          type: string
          nullable: true
          description: "Pct Reajuste Anual"
        dtUltReajuste:
          type: string
          nullable: true
          description: "Dt Ult Reajuste"
        fatorK:
          type: string
          nullable: true
          description: "Fator K"
        rendaMaxPercap:
          type: string
          nullable: true
          description: "Renda Max Percap"
        idadeMin:
          type: string
          nullable: true
          description: "Idade Min"
        idadeMax:
          type: string
          nullable: true
          description: "Idade Max"
        indExigeFilhos:
          type: string
          nullable: true
          description: "Ind Exige Filhos"
        qtdMinFilhos:
          type: string
          nullable: true
          description: "Qtd Min Filhos"
        indExigeEscola:
          type: string
          nullable: true
          description: "Ind Exige Escola"
        indExigeVacina:
          type: string
          nullable: true
          description: "Ind Exige Vacina"
        indExigePrenatal:
          type: string
          nullable: true
          description: "Ind Exige Prenatal"
        indExigeBiometria:
          type: string
          nullable: true
          description: "Ind Exige Biometria"
        dtInclusao:
          type: string
          nullable: true
          description: "Dt Inclusao"
        usrInclusao:
          type: string
          nullable: true
          description: "Usr Inclusao"
        dtUltAlteracao:
          type: string
          nullable: true
          description: "Dt Ult Alteracao"
        usrUltAlteracao:
          type: string
          nullable: true
          description: "Usr Ult Alteracao"
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    ProgramaSocialInput:
      type: object
      properties:
        codPrograma:
          type: string
          nullable: true
          description: "Cod Programa"
        nomePrograma:
          type: string
          nullable: true
          description: "Nome Programa"
        siglaPrograma:
          type: string
          nullable: true
          description: "Sigla Programa"
        tipoPrograma:
          type: string
          nullable: true
          description: "Tipo Programa"
        orgaoResponsavel:
          type: string
          nullable: true
          description: "Orgao Responsavel"
        leiCriacao:
          type: string
          nullable: true
          description: "Lei Criacao"
        dtCriacao:
          type: string
          nullable: true
          description: "Dt Criacao"
        dtEncerramento:
          type: string
          nullable: true
          description: "Dt Encerramento"
        sitPrograma:
          type: string
          nullable: true
          description: "Sit Programa"
        vlrBaseIndividual:
          type: string
          nullable: true
          description: "Vlr Base Individual"
        vlrBaseFamiliar:
          type: string
          nullable: true
          description: "Vlr Base Familiar"
        vlrTetoBenef:
          type: string
          nullable: true
          description: "Vlr Teto Benef"
        vlrPisoBenef:
          type: string
          nullable: true
          description: "Vlr Piso Benef"
        pctReajusteAnual:
          type: string
          nullable: true
          description: "Pct Reajuste Anual"
        dtUltReajuste:
          type: string
          nullable: true
          description: "Dt Ult Reajuste"
        fatorK:
          type: string
          nullable: true
          description: "Fator K"
        rendaMaxPercap:
          type: string
          nullable: true
          description: "Renda Max Percap"
        idadeMin:
          type: string
          nullable: true
          description: "Idade Min"
        idadeMax:
          type: string
          nullable: true
          description: "Idade Max"
        indExigeFilhos:
          type: string
          nullable: true
          description: "Ind Exige Filhos"
        qtdMinFilhos:
          type: string
          nullable: true
          description: "Qtd Min Filhos"
        indExigeEscola:
          type: string
          nullable: true
          description: "Ind Exige Escola"
        indExigeVacina:
          type: string
          nullable: true
          description: "Ind Exige Vacina"
        indExigePrenatal:
          type: string
          nullable: true
          description: "Ind Exige Prenatal"
        indExigeBiometria:
          type: string
          nullable: true
          description: "Ind Exige Biometria"
        dtInclusao:
          type: string
          nullable: true
          description: "Dt Inclusao"
        usrInclusao:
          type: string
          nullable: true
          description: "Usr Inclusao"
        dtUltAlteracao:
          type: string
          nullable: true
          description: "Dt Ult Alteracao"
        usrUltAlteracao:
          type: string
          nullable: true
          description: "Usr Ult Alteracao"
    ProgramaSocialList:
      type: object
      properties:
        data:
          type: array
          items: {$ref: "#/components/schemas/ProgramaSocial"}
        total: {type: integer}
        page: {type: integer}
        limit: {type: integer}
    ProgramaSocialGrpFaixaCalculo:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        rendaInicio:
          type: string
          nullable: true
          description: "Renda Inicio"
        rendaFim:
          type: string
          nullable: true
          description: "Renda Fim"
        fatorMultiplicador:
          type: string
          nullable: true
          description: "Fator Multiplicador"
        vlrAdicional:
          type: string
          nullable: true
          description: "Vlr Adicional"
        indAcumulativo:
          type: string
          nullable: true
          description: "Ind Acumulativo"
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    ProgramaSocialGrpFaixaCalculoInput:
      type: object
      properties:
        rendaInicio:
          type: string
          nullable: true
          description: "Renda Inicio"
        rendaFim:
          type: string
          nullable: true
          description: "Renda Fim"
        fatorMultiplicador:
          type: string
          nullable: true
          description: "Fator Multiplicador"
        vlrAdicional:
          type: string
          nullable: true
          description: "Vlr Adicional"
        indAcumulativo:
          type: string
          nullable: true
          description: "Ind Acumulativo"
    ProgramaSocialGrpFaixaCalculoList:
      type: object
      properties:
        data:
          type: array
          items: {$ref: "#/components/schemas/ProgramaSocialGrpFaixaCalculo"}
        total: {type: integer}
        page: {type: integer}
        limit: {type: integer}
    ProgramaSocialGrpParamRegional:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        codRegiao:
          type: string
          nullable: true
          description: "Cod Regiao"
        fatorRegional:
          type: string
          nullable: true
          description: "Fator Regional"
        vlrComplementoReg:
          type: string
          nullable: true
          description: "Vlr Complemento Reg"
        indAtivoRegiao:
          type: string
          nullable: true
          description: "Ind Ativo Regiao"
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
    ProgramaSocialGrpParamRegionalInput:
      type: object
      properties:
        codRegiao:
          type: string
          nullable: true
          description: "Cod Regiao"
        fatorRegional:
          type: string
          nullable: true
          description: "Fator Regional"
        vlrComplementoReg:
          type: string
          nullable: true
          description: "Vlr Complemento Reg"
        indAtivoRegiao:
          type: string
          nullable: true
          description: "Ind Ativo Regiao"
    ProgramaSocialGrpParamRegionalList:
      type: object
      properties:
        data:
          type: array
          items: {$ref: "#/components/schemas/ProgramaSocialGrpParamRegional"}
        total: {type: integer}
        page: {type: integer}
        limit: {type: integer}
...