# lib/validation/programaSocialGrpFaixaCalculo.ts
import { z } from "zod";

// Validation schema for ProgramaSocialGrpFaixaCalculo — generated by RNC from the UIR.
export const programaSocialGrpFaixaCalculoSchema = z.object({
  occurrence: z.coerce.number().int(),
  rendaInicio: z.coerce.number().nullish(),
  rendaFim: z.coerce.number().nullish(),
  fatorMultiplicador: z.coerce.number().nullish(),
  vlrAdicional: z.coerce.number().nullish(),
  indAcumulativo: z.string().nullish(),
});

export type ProgramaSocialGrpFaixaCalculoInput = z.infer<typeof programaSocialGrpFaixaCalculoSchema>;